requirements.txt
streamlit>=1.32.0
google-api-python-client>=2.120.0
google-generativeai>=0.8.0
pandas>=2.0.0
